banana_api/requirements.txt
# PyTorch versions from your notebook output (Cell 1)
torch==2.7.1
torchvision==0.22.1
timm==1.0.20
# API and utility libraries
fastapi
uvicorn[standard]
pillow